backend/pyproject.toml
[tool.poetry]
name = "sqr-pasynkov-team-backend"
version = "0.1.0"
description = ""
authors = ["Sergei Pasynkov <the.s.pasynkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.1"
pydantic = "^2.7.0"
sqlalchemy = "^2.0.29"
unicorn = "^2.0.1.post1"
pytest = "^8.1.1"
bandit = "^1.7.8"
flake8 = "^7.0.0"
pytest-cov = "^5.0.0"
uvicorn = "^0.29.0"
bcrypt = "^4.1.2"
pyjwt = "^2.8.0"
pydantic-settings = "^2.2.1"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
